+++ gpjax/__init__.py
@@ -20,8 +20,8 @@ from .kernels import (
 from .likelihoods import Bernoulli, Gaussian
 from .mean_functions import Constant, Zero
 from .parameters import copy_dict_structure, initialise, transform
-from .sparse_gps import StochasticVI
 from .types import Dataset
-from .variational import VariationalGaussian, WhitenedVariationalGaussian
+from .variational_inference import StochasticVI
+from .variational_families import VariationalGaussian, WhitenedVariationalGaussian
 
 __version__ = "0.4.6"
