+++ tests/test_variational_inference.py
